comment_service/pyproject.toml
[tool.poetry]
name = "comment-service"
version = "0.1.0"
description = ""
authors = ["Yevgeniy Kap <kapustaevg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.2"
pydantic-settings = "^2.5.2"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
fastapi-cache2 = "^0.2.2"
redis = "^5.0.8"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
